Splits args and config modules
