+++ src/config.rs
@@ -1,39 +1,26 @@
-use clap::Parser;
+use crate::args::Args;
 use color_eyre::Result;
 use serde::{Deserialize, Serialize};
 use std::default::Default;
 use std::path::PathBuf;
 
 // single struct to hold configuration and arguments
-#[derive(Parser, Debug, Serialize, Deserialize, Clone)]
-#[command(version, about, long_about)]
+#[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Config {
-    #[arg(short, long)]
-    pub no_color: bool,
-    #[arg(short, long)]
     pub enabled_string: Option<String>,
-    #[arg(long)]
     pub enabled_color: Option<String>,
-    #[arg(long)]
     pub enabled_style: Option<String>, // emphasys
-    #[arg(short, long)]
     pub disabled_string: Option<String>,
-    #[arg(long)]
     pub disabled_color: Option<String>,
-    #[arg(long)]
     pub disabled_style: Option<String>,
-    #[arg(short, long)]
     pub config_path: Option<PathBuf>,
-    #[arg(short, long)]
-    pub lookup: bool,
-    #[arg(short = 'p', long)]
+    pub lookup: Option<bool>,
     pub lookup_provider: Option<String>,
 }
 
 impl Default for Config {
     fn default() -> Self {
         Self {
-            no_color: false,
             enabled_string: Some("enabled".to_string()),
             enabled_color: Some("green".to_string()),
             enabled_style: None,
@@ -41,7 +28,7 @@ impl Default for Config {
             disabled_color: Some("red".to_string()),
             disabled_style: None,
             config_path: None,
-            lookup: false,
+            lookup: None,
             lookup_provider: None,
         }
     }
@@ -58,14 +45,10 @@ impl Config {
         Ok(config)
     }
 
-    fn parse_args() -> Self {
-        Config::parse()
-    }
-
     pub fn get() -> Self {
         let mut config = Config::default();
         // parse cli arguments
-        let args = Config::parse_args();
+        let args = Args::parse_args();
 
         // load config file, if no path was specified in args use default path
         // if no file is found at the default path, use default values and save new config file
@@ -86,9 +69,8 @@ impl Config {
         if args.disabled_color.is_some() {
             config.disabled_color = args.disabled_color;
         }
-
         if args.lookup {
-            config.lookup = args.lookup;
+            config.lookup = Some(args.lookup);
         }
 
         config
@@ -98,27 +80,6 @@ impl Config {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn parse_arguments() {
-        let args = vec![
-            "vpn_status",
-            "--enabled-string",
-            "active",
-            "--enabled-color",
-            "blue",
-            "--disabled-string",
-            "inactive",
-            "--disabled-color",
-            "yellow",
-        ];
-
-        let config = Config::parse_from(args);
-        assert_eq!(config.enabled_string, Some("active".to_string()));
-        assert_eq!(config.enabled_color, Some("blue".to_string()));
-        assert_eq!(config.disabled_string, Some("inactive".to_string()));
-        assert_eq!(config.disabled_color, Some("yellow".to_string()));
-    }
 
     #[test]
     fn load_config() {
